Create scripts to create and update database.
Created sql files that contain the schema of GEOmetadb and bioqc.
Created a script that reads the sqlite database and writes it into
postgreSQL. Created an sql files that updates primary keys and
foreign key constraints in GEOmetadb and fixes inconsistencies.
Created an Rmd script, that checks the consistency between
sqlite and postgres.